Change `fromServiceInfo` return value to ServiceInstance
There was an issue with Javadoc and Lombok generated builders because
javadoc does not process annotations first and thus `fromServiceInfo`'s
return value of `ServiceInstanceBuilder` caused javadoc generation
failures.

This changes the method to return an actual ServiceInstance and in
order to still allow for setting the instanceId and status, this commit
adds `@With` annotations for those fields.