Remove useless Mockito spies from EurekaRegistryClientTest (#154)
Remove the Mockito spies from EurekaRegistryClient in the
EurekaRegistryClientTest. It turns out they were not actually doing
anything at all and thus were not changing the WaitStrategy in the retry
tests. The reason is that EurekaRegistryClient's constructor creates
a KiwiRetryer once, and uses the default wait strategy (which in our
case is an exponential strategy). So, after the EurekaRegistryClient has
been created, getWaitStrategy() is never called again and thus the
Mockito stubbing had no effect whatsoever!

Change the top-level setUp so that it creates a "real" instance of
EurekaRegistryClient, and change the setUp in the WithRetry nested test
class so that it overrides the getWaitStrategy method to return a
"no wait" strategy. This makes the retry tests way faster, e.g. about
4-5 millis instead of about 600 millis (the first test in WithRetry
always incurs the several hundred millisecond hit to initialize the
Mockito mock of the EurekaRestClient). But overall this shaves over a
second off this entire test class. Totally unscientific results on this
Linux workstation reduced the EurekaRegistryClientTest execution time
from consistently taking over 2 seconds to less than a second, sometimes
as low as 800 millis and other times in the mid 900 millis.